+++ app/src/main/AndroidManifest.xml
@@ -13,7 +13,7 @@
         android:theme="@style/Theme.Organeco"
         tools:targetApi="31">
         <activity
-            android:name=".view.activity.CekSubsidi.CekSubsidiActivity"
+            android:name=".view.activity.checksubsidies.CekSubsidiActivity"
             android:exported="false" />
 
         <activity
