+++ app/src/main/res/layout/activity_cek_subsidi.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".view.activity.CekSubsidi.CekSubsidiActivity">
+    tools:context=".view.activity.checksubsidies.CekSubsidiActivity">
 
     <androidx.appcompat.widget.Toolbar
         android:id="@+id/toolbar"
